requirements.txt
pika
redis
python-dotenv
pytest
coverage